mesh: trace image points onto mesh

+++ libs/MVS/Image.cpp
@@ -150,7 +150,7 @@ float Image::ResizeImage(unsigned nMaxResolution)
 	height = (uint32_t)scaledSize.height;
 	if (!image.empty())
 		cv::resize(image, image, scaledSize, 0, 0, cv::INTER_AREA);
-	return scale;
+	return static_cast<float>(scale);
 } // ResizeImage
 /*----------------------------------------------------------------*/
